Fix oob error (#2)
If last position of an alignment had mutated into a stop codon - which happened on a small number of sequences where the last codon had mutated and was no longer complete in the alignment, and the second last codon had mutated into a stop codon. This caused the amino acid alignment to be one residue short in the query sequence than the reference sequence since terminal stop codons are trimmed off by the biojava alignment package.

Also updated Dockerfile so builds now succeed.